+++ multi/imxrt-multi/imxrt-multi.c
@@ -276,7 +276,7 @@ static int createDevFiles(void)
 static void multi_thread(void *arg)
 {
 	msg_t msg;
-	unsigned int rid;
+	unsigned long int rid;
 
 	while (1) {
 		while (msgRecv(multi_port, &msg, &rid) < 0)
@@ -319,7 +319,7 @@ static void multi_thread(void *arg)
 static void uart_thread(void *arg)
 {
 	msg_t msg;
-	unsigned int rid;
+	unsigned long int rid;
 
 	while (1) {
 		while (msgRecv(common.uart_port, &msg, &rid) < 0)
